+++ alacritty/src/display/content.rs
@@ -190,6 +190,7 @@ pub struct RenderableCell {
     pub fg: Rgb,
     pub bg: Rgb,
     pub bg_alpha: f32,
+    pub underline_color: Rgb,
     pub flags: Flags,
 }
 
@@ -251,12 +252,22 @@ impl RenderableCell {
             Self::compute_cell_rgb(&mut fg, &mut bg, &mut bg_alpha, config_fg, config_bg);
         }
 
+        let (zerowidth, underline_color) = if let Some(extra) = cell.extra() {
+            (Some(extra.zerowidth.clone()), extra.underline_color)
+        } else {
+            (None, None)
+        };
+
+        let underline_rgb =
+            underline_color.map_or(fg, |color| Self::compute_fg_rgb(content, color, cell.flags));
+
         // Convert cell point to viewport position.
         let cell_point = cell.point;
         let point = display::point_to_viewport(display_offset, cell_point).unwrap();
 
         RenderableCell {
-            zerowidth: cell.zerowidth().map(|zerowidth| zerowidth.to_vec()),
+            zerowidth,
+            underline_color: underline_rgb,
             flags: cell.flags,
             character,
             bg_alpha,